initial 1.5 upload

+++ xbmc/FileSystem/CacheStrategy.cpp
@@ -190,7 +190,7 @@ int CSimpleFileCache::ReadFromCache(char *pBuffer, size_t iMaxSize)
 		return m_bEndOfInput?CACHE_RC_EOF : CACHE_RC_WOULD_BLOCK;
 	}
 
-	if (iMaxSize > iAvailable)
+	if ((int64_t) iMaxSize > iAvailable)
 		iMaxSize = (size_t)iAvailable;
 
 	DWORD iRead = 0;
